refactor: split receiver into actor and implementation
