+++ broker/src/main/java/io/camunda/zeebe/broker/transport/partitionapi/InterPartitionCommandReceiverImpl.java
@@ -7,12 +7,8 @@
  */
 package io.camunda.zeebe.broker.transport.partitionapi;
 
-import static io.camunda.zeebe.broker.transport.partitionapi.InterPartitionCommandSenderImpl.TOPIC_PREFIX;
-
 import io.atomix.cluster.MemberId;
-import io.atomix.cluster.messaging.ClusterCommunicationService;
 import io.camunda.zeebe.broker.Loggers;
-import io.camunda.zeebe.broker.system.monitoring.DiskSpaceUsageListener;
 import io.camunda.zeebe.clustering.management.InterPartitionMessageDecoder;
 import io.camunda.zeebe.clustering.management.MessageHeaderDecoder;
 import io.camunda.zeebe.logstreams.log.LogStreamRecordWriter;
@@ -20,80 +16,22 @@ import io.camunda.zeebe.protocol.impl.record.RecordMetadata;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.Intent;
-import io.camunda.zeebe.scheduler.Actor;
 import io.camunda.zeebe.util.buffer.BufferWriter;
 import io.camunda.zeebe.util.buffer.DirectBufferWriter;
-import java.util.Map;
 import org.agrona.concurrent.UnsafeBuffer;
 import org.slf4j.Logger;
 
-/**
- * Receives messages send by @{@link InterPartitionCommandSenderImpl} and tries to write them as
- * commands to the partition's log stream. Failure to write to the log stream, for example because
- * no disk space is available, the logstream rejected the write operation or message decoding
- * failure, are ignored. The sender is responsible for recognizing failures and retrying.
- */
-public final class InterPartitionCommandReceiver extends Actor implements DiskSpaceUsageListener {
+public final class InterPartitionCommandReceiverImpl {
   private static final Logger LOG = Loggers.TRANSPORT_LOGGER;
-  private final String actorName;
-  private final ClusterCommunicationService communicationService;
-  private final LogStreamRecordWriter logStreamWriter;
-  private final int partitionId;
   private final Decoder decoder = new Decoder();
+  private final LogStreamRecordWriter logStreamWriter;
   private boolean diskSpaceAvailable = true;
 
-  public InterPartitionCommandReceiver(
-      final int nodeId,
-      final int partitionId,
-      final ClusterCommunicationService communicationService,
-      final LogStreamRecordWriter logStreamWriter) {
-    this.partitionId = partitionId;
-    this.communicationService = communicationService;
+  public InterPartitionCommandReceiverImpl(final LogStreamRecordWriter logStreamWriter) {
     this.logStreamWriter = logStreamWriter;
-    actorName = buildActorName(nodeId, getClass().getSimpleName(), partitionId);
   }
 
-  @Override
-  protected Map<String, String> createContext() {
-    final var context = super.createContext();
-    context.put(ACTOR_PROP_PARTITION_ID, Integer.toString(partitionId));
-    return context;
-  }
-
-  @Override
-  public String getName() {
-    return actorName;
-  }
-
-  @Override
-  protected void onActorStarting() {
-    communicationService.subscribe(TOPIC_PREFIX + partitionId, this::tryHandleMessage, actor::run);
-  }
-
-  @Override
-  protected void onActorClosing() {
-    communicationService.unsubscribe(TOPIC_PREFIX + partitionId);
-  }
-
-  @Override
-  public void onDiskSpaceNotAvailable() {
-    actor.run(() -> diskSpaceAvailable = false);
-  }
-
-  @Override
-  public void onDiskSpaceAvailable() {
-    actor.run(() -> diskSpaceAvailable = true);
-  }
-
-  private void tryHandleMessage(final MemberId memberId, final byte[] message) {
-    try {
-      handleMessage(memberId, message);
-    } catch (final RuntimeException e) {
-      LOG.error("Error while handling message", e);
-    }
-  }
-
-  private void handleMessage(final MemberId memberId, final byte[] message) {
+  public void handleMessage(final MemberId memberId, final byte[] message) {
     LOG.trace("Received message from {}", memberId);
 
     final var decoded = decoder.decodeMessage(message);
@@ -119,6 +57,10 @@ public final class InterPartitionCommandReceiver extends Actor implements DiskSp
     }
   }
 
+  public void setDiskSpaceAvailable(final boolean available) {
+    diskSpaceAvailable = available;
+  }
+
   private static final class Decoder {
     private final UnsafeBuffer messageBuffer = new UnsafeBuffer();
     private final RecordMetadata recordMetadata = new RecordMetadata();
@@ -126,15 +68,14 @@ public final class InterPartitionCommandReceiver extends Actor implements DiskSp
     private final MessageHeaderDecoder headerDecoder = new MessageHeaderDecoder();
     private final DirectBufferWriter commandBuffer = new DirectBufferWriter();
 
-    DecodedMessage decodeMessage(final byte[] message) {
+    Decoder.DecodedMessage decodeMessage(final byte[] message) {
       messageBuffer.wrap(message);
       messageDecoder.wrapAndApplyHeader(messageBuffer, 0, headerDecoder);
 
       final var valueType = ValueType.get(messageDecoder.valueType());
       final var intent = Intent.fromProtocolValue(valueType, messageDecoder.intent());
 
-      // rebuild the record metadata first, all messages contain commands for the current
-      // partitionId
+      // rebuild the record metadata first, all messages must contain commands
       recordMetadata.reset().recordType(RecordType.COMMAND).valueType(valueType).intent(intent);
 
       // wrap the command buffer around the rest of the message
@@ -145,7 +86,7 @@ public final class InterPartitionCommandReceiver extends Actor implements DiskSp
       final var commandLength = messageDecoder.commandLength();
       commandBuffer.wrap(messageBuffer, commandOffset, commandLength);
 
-      return new DecodedMessage(recordMetadata, commandBuffer);
+      return new Decoder.DecodedMessage(recordMetadata, commandBuffer);
     }
 
     record DecodedMessage(RecordMetadata metadata, BufferWriter command) {}
